Refactoring, devices now produce messages to be send by output protocol
