+++ gateway/lora_gateway/include/gateway/device/Device.hpp
@@ -1,12 +1,12 @@
 #pragma once
 
-#include <gateway/structures/DeviceMessage.hpp>
+#include <gateway/input_protocol/InputProtocolMessage.hpp>
 #include <gateway/common_tools/EnumTools.hpp>
-
-#include <boost/json.hpp>
+#include <gateway/device/Message.hpp>
 
 #include <cstdint>
 #include <string>
+#include <memory>
 
 
 
@@ -17,56 +17,22 @@ namespace gateway::device {
  */
 class Device {
 public:
-	Device(structures::EDeviceType deviceType, uint32_t deviceNumber): deviceType_(deviceType),
+	Device(input_protocol::EDeviceType deviceType, uint32_t deviceNumber): deviceType_(deviceType),
 																	   deviceNumber_(deviceNumber) {};
 
 	/**
 	 * @brief Parse given message
-	 * @param message raw message to be parsed
+	 * @param message input protocol raw binary message to be parsed
 	 * @return true if message was parsed successfully, false otherwise
 	 */
-	bool parseMessage(const structures::DeviceMessage &message);
+	virtual std::shared_ptr<Message> parseMessage(const input_protocol::InputProtocolMessage &message) = 0;
 
-	/**
-	 * @brief Get the Csv header of this instances device type
-	 * @return csv header
-	 */
-	[[nodiscard]] virtual std::string getCsvHeader() const = 0;
-
-	/**
-	 * @brief Get the Csv entry of this instances device type
-	 * @return csv entry
-	 */
-	[[nodiscard]]virtual std::string getCsvEntry() const = 0;
-
-	/**
-	 * @brief Get data of the device to publish
-	 * @return
-	 */
-	[[nodiscard]]virtual boost::json::object getOutputProtocolEntry() const = 0;
-
-	/**
-	 * @brief Get the type of the device
-	 * @return device type
-	 */
-	[[nodiscard]] structures::EDeviceType getDeviceType() const;
-
-	/**
-	 * @brief Get the number of the device
-	 * @return device number
-	 */
 	[[nodiscard]] uint32_t getDeviceNumber() const;
 
-	/**
-	 * @brief Had last message wrong checksum?
-	 * @return true for yes, false for no
-	 */
-	[[nodiscard]] bool isWrongChecksum() const;
+	[[nodiscard]] input_protocol::EDeviceType getDeviceType() const;
 
-	/**
-	 * @brief Was message before last lost?
-	 * @return true for yes, false for no
-	 */
+	[[nodiscard]] bool isCorrectChecksum() const;
+
 	[[nodiscard]] bool isMessageLost() const;
 
 protected:
@@ -79,18 +45,23 @@ protected:
 	/// Maximum index of the message
 	static constexpr uint32_t MAX_MESSAGE_INDEX { 15 };
 	/// Type of the device
-	structures::EDeviceType deviceType_;
+	input_protocol::EDeviceType deviceType_;
 	/// ID of the device
 	uint32_t deviceNumber_ { 0 };
+	/// Most current message
 	/// Stamp of the last received message
 	int64_t lastStamp_ { -1 };
 	/// Is the checksum on last parsed message wrong
-	bool isWrongChecksum_ { false };
+	bool isCorrectChecksum_ { false };
 	/// Is the message lost
 	bool isMessageLost_ { false };
-	uint64_t currentTimestampMs_{0};
-	uint8_t flags_{0};
 
-	virtual bool parseData(const structures::DeviceMessage &message) = 0;
+	bool parseCommonHeader(const input_protocol::InputProtocolMessage &message);
+
+	void validateMessageStamp(uint8_t currentStamp);
+
+	void validateChecksum(const input_protocol::InputProtocolMessage &message);
+
+	virtual bool parseDeviceSpecificFlags(const uint8_t &flags) = 0;
 };
 }
