+++ gateway/lora_gateway/source/gateway/device/Device.cpp
@@ -1,5 +1,4 @@
 #include <gateway/device/Device.hpp>
-#include <gateway/common_tools/TimeTools.hpp>
 #include <gateway/common_tools/MessageTools.hpp>
 #include <gateway/logger/Logger.hpp>
 
@@ -7,8 +6,8 @@
 
 namespace gateway::device {
 
-bool Device::parseMessage(const structures::DeviceMessage &message) {
-	if(common_tools::EnumTools::valueToEnum<structures::EDeviceType, uint32_t>(message.deviceType) != deviceType_) {
+bool Device::parseCommonHeader(const input_protocol::InputProtocolMessage &message) {
+	if(common_tools::EnumTools::valueToEnum<input_protocol::EDeviceType, uint32_t>(message.deviceType) != deviceType_) {
 		logger::Logger::logError("Incorrect device type");
 		return false;
 	}
@@ -17,50 +16,17 @@ bool Device::parseMessage(const structures::DeviceMessage &message) {
 		return false;
 	}
 
-	isWrongChecksum_ = false;
-	isMessageLost_ = false;
-	currentTimestampMs_ = common_tools::TimeTools::getUnixTimestampMs();
-
-	if(message.flags & WRONG_CHECKSUM_MASK) {
-		logger::Logger::logError("Wrong check sum on 433MHZ RF radio");
-		isWrongChecksum_ = true;
-	}
-
-	if(message.flags & MESSAGE_LOST_MASK) {
-		logger::Logger::logError("Message got lost on 433MHZ RF radio");
-		isMessageLost_ = true;
-	}
+	validateChecksum(message);
 
 	uint8_t stamp = (message.flags & STAMP_MASK);
+	validateMessageStamp(stamp);
 
-	flags_ = message.flags;
+	parseDeviceSpecificFlags(message.flags);
 
-	if(lastStamp_ != -1) {
-		uint32_t nextStamp = lastStamp_ + 1;
-		if(nextStamp > MAX_MESSAGE_INDEX) {
-			nextStamp = 0;
-		}
-		if(nextStamp != stamp) {
-			logger::Logger::logError("Message got lost on LoRa radio");
-			isMessageLost_ = true;
-		}
-	}
-	lastStamp_ = stamp;
-
-	uint32_t checkSum = common_tools::MessageTools::calculateChecksum((uint8_t *)&message,
-																	  sizeof(structures::DeviceMessage) -
-																	  sizeof(uint32_t));
-	if(checkSum != message.checkSum) {
-		isWrongChecksum_ = true;
-		logger::Logger::logError("Wrong checksum");
-	}
-
-	parseData(message);
-
-	return !isWrongChecksum_;
+	return isCorrectChecksum_;
 }
 
-structures::EDeviceType Device::getDeviceType() const {
+input_protocol::EDeviceType Device::getDeviceType() const {
 	return deviceType_;
 }
 
@@ -68,11 +34,33 @@ uint32_t Device::getDeviceNumber() const {
 	return deviceNumber_;
 }
 
-bool Device::isWrongChecksum() const {
-	return isWrongChecksum_;
+bool Device::isCorrectChecksum() const {
+	return isCorrectChecksum_;
 }
 
 bool Device::isMessageLost() const {
 	return isMessageLost_;
 }
+
+void Device::validateMessageStamp(uint8_t currentStamp) {
+	if(lastStamp_ != -1) {
+		uint32_t nextStamp = lastStamp_ + 1;
+		if(nextStamp > MAX_MESSAGE_INDEX) {
+			nextStamp = 0;
+		}
+		if(nextStamp != currentStamp) {
+			logger::Logger::logError("Message got lost on LoRa radio");
+			isMessageLost_ = true;
+		}
+	}
+	lastStamp_ = currentStamp;
+}
+
+void Device::validateChecksum(const input_protocol::InputProtocolMessage &message) {
+	uint32_t checkSum = common_tools::MessageTools::calculateChecksum((uint8_t *)&message,
+																	  sizeof(input_protocol::InputProtocolMessage) -
+																	  sizeof(uint32_t));
+	isCorrectChecksum_ = (checkSum == message.checkSum);
+}
+
 }
