--- gateway/lora_gateway/source/gateway/endpoint/Endpoint.cpp
@@ -1,7 +1,0 @@
-#include <gateway/endpoint/Endpoint.hpp>
-
-
-
-namespace gateway::endpoint {
-
-}
