--- gateway/lora_gateway/source/gateway/endpoint/EndpointFactory.cpp
@@ -1,19 +1,0 @@
-#include <gateway/endpoint/EndpointFactory.hpp>
-#include <gateway/endpoint/lora/LoraEndpoint.hpp>
-#include <gateway/endpoint/generator/GeneratorEndpoint.hpp>
-
-
-
-namespace gateway::endpoint {
-
-std::shared_ptr<Endpoint> EndpointFactory::createEndpoint(const std::shared_ptr<structures::GlobalContext> &context) {
-	switch(context->settings->getDeviceType()) {
-		case settings::EDeviceCommunicationType::E_LORA:
-			return std::make_shared<lora::LoraEndpoint>(context);
-		case settings::EDeviceCommunicationType::E_GENERATOR:
-			return std::make_shared<generator::GeneratorEndpoint>(context);
-		default:
-			return nullptr;
-	}
-}
-}
