+++ gateway/lora_gateway/source/gateway/input_protocol/Endpoint.cpp
@@ -1,0 +1,7 @@
+#include <gateway/input_protocol/Endpoint.hpp>
+
+
+
+namespace gateway::endpoint {
+
+}
